requirements.txt
bs4
requests
scikit-learn
umap-learn
keras
tensorflow
python-unsplash
pillow
shutil
io
pytesseract